+++ src/ServiceProvider.php
@@ -5,6 +5,7 @@ namespace Calisero\LaravelSms;
 use Calisero\LaravelSms\Console\Commands\CheckVerificationCommand;
 use Calisero\LaravelSms\Console\Commands\SendTestSmsCommand;
 use Calisero\LaravelSms\Console\Commands\SendVerificationCommand;
+use Calisero\LaravelSms\Console\Commands\StatusSmsCommand;
 use Calisero\LaravelSms\Contracts\SmsClient as SmsClientContract;
 use Calisero\LaravelSms\Notification\SmsChannel;
 use Illuminate\Notifications\ChannelManager;
@@ -61,6 +62,7 @@ class ServiceProvider extends BaseServiceProvider
                 SendTestSmsCommand::class,
                 SendVerificationCommand::class,
                 CheckVerificationCommand::class,
+                StatusSmsCommand::class,
             ]);
         }
     }